requirements.txt
python-telegram-bot==12.8
kaleido==0.2.0
plotly==4.9.0
PyJWT==1.7.1
pandas==1.5.0
requests==2.24.0
